# server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/system/audit-logs": {
      "get": {
        "operationId": "AuditLogsController_findAllSystem",
        "parameters": [
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "全体監査ログ一覧（system_admin）",
        "tags": [
          "audit-logs"
        ]
      }
    },
    "/api/v1/admin/audit-logs": {
      "get": {
        "operationId": "AuditLogsController_findAllTenant",
        "parameters": [
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "テナント内監査ログ一覧（tenant_admin）",
        "tags": [
          "audit-logs"
        ]
      }
    },
...
    "/api/v1/attendance/clock-in": {
      "post": {
        "operationId": "AttendanceController_clockIn",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "出勤打刻",
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/attendance/clock-out": {
      "post": {
        "operationId": "AttendanceController_clockOut",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "退勤打刻",
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/attendance/break/start": {
      "post": {
        "operationId": "AttendanceController_breakStart",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreakRecordResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "休憩開始",
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/attendance/break/end": {
      "post": {
        "operationId": "AttendanceController_breakEnd",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreakRecordResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "休憩終了",
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/attendance/today": {
      "get": {
        "operationId": "AttendanceController_getToday",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "本日の勤怠状態取得",
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/attendance": {
      "get": {
        "operationId": "AttendanceController_findMine",
        "parameters": [
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "自身の勤怠一覧（月指定）",
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/attendance/summary": {
      "get": {
        "operationId": "AttendanceController_getMySummary",
        "parameters": [
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceSummaryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "自身の月次集計",
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/admin/attendance": {
      "get": {
        "operationId": "AttendanceController_findAllAdmin",
        "parameters": [
          {
            "name": "employee_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "テナント内全従業員の勤怠一覧（tenant_admin）",
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/admin/attendance/summary": {
      "get": {
        "operationId": "AttendanceController_getAdminSummary",
        "parameters": [
          {
            "name": "employee_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceSummaryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "テナント内月次集計（tenant_admin）",
        "tags": [
          "attendance"
        ]
      }
    },
...
  },
  "info": {
    "title": "勤怠管理システム API",
    "description": "マルチテナント勤怠管理システムのREST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AuditLogActorResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email"
        ]
      },
      "AuditLogResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "object",
            "nullable": true
          },
          "actorId": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "targetType": {
            "type": "string"
          },
          "targetId": {
            "type": "object",
            "nullable": true
          },
          "detail": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "actor": {
            "$ref": "#/components/schemas/AuditLogActorResponseDto"
          }
        },
        "required": [
          "id",
          "tenantId",
          "actorId",
          "action",
          "targetType",
          "targetId",
          "detail",
          "createdAt",
          "actor"
        ]
      },
      "AuditLogListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLogResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
...
      "BreakRecordResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "attendanceId": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "attendanceId",
          "startTime",
          "endTime"
        ]
      },
      "AttendanceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "employeeId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "clockIn": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "clockOut": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "working",
              "completed",
              "holiday",
              "absent"
            ]
          },
          "isOvernight": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "breakRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BreakRecordResponseDto"
            }
          }
        },
        "required": [
          "id",
          "tenantId",
          "employeeId",
          "date",
          "clockIn",
          "clockOut",
          "status",
          "isOvernight",
          "createdAt",
          "updatedAt",
          "breakRecords"
        ]
      },
      "AttendanceEmployeeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "employeeNumber": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "employeeNumber"
        ]
      },
      "AttendanceListItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "employeeId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "clockIn": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "clockOut": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "working",
              "completed",
              "holiday",
              "absent"
            ]
          },
          "isOvernight": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "breakRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BreakRecordResponseDto"
            }
          },
          "employee": {
            "$ref": "#/components/schemas/AttendanceEmployeeResponseDto"
          }
        },
        "required": [
          "id",
          "tenantId",
          "employeeId",
          "date",
          "clockIn",
          "clockOut",
          "status",
          "isOvernight",
          "createdAt",
          "updatedAt",
          "breakRecords",
          "employee"
        ]
      },
      "AttendanceListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttendanceListItemResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
      "AttendanceSummaryResponseDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "number"
          },
          "month": {
            "type": "number"
          },
          "totalWorkMinutes": {
            "type": "number"
          },
          "totalBreakMinutes": {
            "type": "number"
          },
          "totalOvertimeMinutes": {
            "type": "number"
          },
          "totalNightMinutes": {
            "type": "number"
          },
          "presentDays": {
            "type": "number"
          },
          "absentDays": {
            "type": "number"
          },
          "holidayDays": {
            "type": "number"
          }
        },
        "required": [
          "year",
          "month",
          "totalWorkMinutes",
          "totalBreakMinutes",
          "totalOvertimeMinutes",
          "totalNightMinutes",
          "presentDays",
          "absentDays",
          "holidayDays"
        ]
      },
...
    }
  }
}